pyproject.toml
[project]
name = "bitcoin-facebook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
    "playwright>=1.54.0",
    "psycopg2-binary>=2.9.10",
    "pytest-playwright>=0.7.0",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]
